+++ src/data/explorer.py
@@ -1,4 +1,5 @@
-from data.init import curs
+from data.init import curs, IntegrityError
+from error import Missing, Duplicate
 from model.explorer import Explorer
 
 curs.execute(
@@ -13,7 +14,8 @@ curs.execute(
 
 
 def row_to_model(row: tuple) -> Explorer:
-    return Explorer(name=row[0], country=row[1], description=row[2])
+    name, country, description = row
+    return Explorer(name=name, country=country, description=description)
 
 
 def model_to_dict(explorer: Explorer) -> dict:
@@ -24,7 +26,11 @@ def get_one(name: str) -> Explorer:
     qry = "select * from explorer where name = :name"
     params = {"name": name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
 
 def get_all() -> list[Explorer]:
@@ -34,15 +40,22 @@ def get_all() -> list[Explorer]:
 
 
 def create(explorer: Explorer) -> Explorer:
+    if not explorer:
+        return None
     qry = """insert into explorer (name, country, description)
         values (:name, :country, :description)
     """
     params = model_to_dict(explorer)
-    _ = curs.execute(qry, params)
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise Duplicate(msg=f"Explorer {explorer.name} already exists")
     return get_one(explorer.name)
 
 
 def modify(name: str, explorer: Explorer) -> Explorer:
+    if not (name and explorer):
+        return None
     qry = """update explorer
         set country = :country
             name = :name
@@ -51,13 +64,18 @@ def modify(name: str, explorer: Explorer) -> Explorer:
     """
     params = model_to_dict(explorer)
     params["name_orig"] = explorer.name
-    _ = curs.execute(qry, params)
-    explorer2 = get_one(explorer.name)
-    return explorer2
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str):
+    if not name:
+        return False
     qry = "delete from explorer where name = :name"
-    params = {"name": explorer.name}
-    res = curs.execute(qry, params)
-    return bool(res)
+    params = {"name": name}
+    curs.execute(qry, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Explorer {name} not found")
